Reorganize ui notifications service
