+++ ui/main/src/app/app.component.ts
@@ -10,7 +10,7 @@
 import {Component, HostListener, TemplateRef, ViewChild} from '@angular/core';
 import {LogOption, LoggerService as logger} from 'app/services/logs/LoggerService';
 import {RemoteLoggerService} from 'app/services/logs/RemoteLoggerService';
-import {SoundNotificationService} from 'app/business/services/notifications/sound-notification.service';
+import {SoundNotificationService} from '@ofServices/notifications/SoundNotificationService';
 import {OpfabEventStreamService} from './business/services/events/opfabEventStream.service';
 import {RouterNavigationService} from 'app/router/router-navigation.service';
 import {PageType, RouterStore} from './business/store/router.store';
