+++ ui/main/src/app/business/application-loader.ts
@@ -28,23 +28,23 @@ import {CardService} from './services/card/card.service';
 import {SupervisedEntitiesService} from './services/users/supervised-entities.service';
 import {SelectedCardLoaderService} from './services/card/selectedCardLoader.service';
 import {HandlebarsService} from './services/card/handlebars.service';
-import {ExternalDevicesService} from './services/notifications/external-devices.service';
+import {ExternalDevicesService} from '../services/notifications/ExternalDevicesService';
 import {TemplateCssService} from './services/card/template-css.service';
 import {SettingsService} from './services/users/settings.service';
 import {RouterService} from './services/router.service';
 import {OpfabStore} from './store/opfabStore';
 import {ApplicationUpdateService} from './services/events/application-update.service';
-import {SystemNotificationService} from './services/notifications/system-notification.service';
+import {SystemNotificationService} from '../services/notifications/SystemNotificationService';
 import {ApplicationLoadingComponent} from './application-loading-component';
 import {ServerResponseStatus} from './server/serverResponse';
 import {Utilities} from './common/utilities';
 import {ModalService} from './services/modal.service';
 import {SessionManagerService} from './services/session-manager.service';
-import {SoundNotificationService} from './services/notifications/sound-notification.service';
+import {SoundNotificationService} from '../services/notifications/SoundNotificationService';
 import {I18n} from '@ofModel/i18n.model';
 import {ProcessStatesMultiSelectOptionsService} from './services/process-states-multi-select-options.service';
 import {RealtimeDomainService} from './services/realtime-domain.service';
-import {NotificationDecision} from './services/notifications/notification-decision';
+import {NotificationDecision} from '../services/notifications/NotificationDecision';
 import {CardTemplateGateway} from './templateGateway/cardTemplateGateway';
 import {UserCardTemplateGateway} from './templateGateway/userCardTemplateGateway';
