+++ ui/main/src/app/modules/core/application-loading/application-loading.component.ts
@@ -31,11 +31,11 @@ import {AdminProcessServer} from 'app/business/server/adminprocess.server';
 import {BusinessDataServer} from 'app/business/server/businessData.server';
 import {CardServer} from 'app/business/server/card.server';
 import {SupervisedEntitiesServer} from 'app/business/server/supervised-entities.server';
-import {ExternalDevicesServer} from 'app/business/server/external-devices.server';
+import {ExternalDevicesServer} from '@ofServices/notifications/server/ExternalDevicesServer';
 import {TemplateCssServer} from '../../../business/server/template-css.server';
 import {SettingsServer} from '../../../business/server/settings.server';
 import {ModalServer} from 'app/business/server/modal.server';
-import {SoundServer} from 'app/business/server/sound.server';
+import {SoundServer} from '@ofServices/notifications/server/SoundServer';
 import {NgIf} from '@angular/common';
 import {LoadingInProgressComponent} from './loading-in-progress/loading-in-progress.component';
 import {LoginComponent} from './login/login.component';
