+++ ui/main/src/app/services/notifications/model/ExternalDevices.ts
