+++ ui/main/src/main.ts
@@ -45,12 +45,12 @@ import {SettingsServer} from './app/business/server/settings.server';
 import {AngularSettingsServer} from './app/server/angularSettings.server';
 import {OpfabEventStreamServer} from './app/business/server/opfabEventStream.server';
 import {AngularOpfabEventStreamServer} from './app/server/angularOpfabEventStream.server';
-import {ExternalDevicesServer} from './app/business/server/external-devices.server';
-import {AngularExternalDevicesServer} from './app/server/angularExternalDevices.server';
+import {ExternalDevicesServer} from './app/services/notifications/server/ExternalDevicesServer';
+import {AngularExternalDevicesServer} from './app/services/notifications/server/AngularExternalDevicesServer';
 import {CardServer} from './app/business/server/card.server';
 import {AngularCardServer} from './app/server/angularCard.server';
-import {SoundServer} from './app/business/server/sound.server';
-import {AngularSoundServer} from './app/server/angularSound.server';
+import {SoundServer} from './app/services/notifications/server/SoundServer';
+import {AngularSoundServer} from './app/services/notifications/server/AngularSoundServer';
 import {TranslationService} from './app/business/services/translation/translation.service';
 import {AngularTranslationService} from '@ofServices/angularTranslationService';
 import {ModalServer} from './app/business/server/modal.server';
